eggimagemenubutton: Copy and un-deprecate GtkImageMenuitem
This was deprecated without reference to a bug or reasoning,
and we need it for our importer drop down button. We want
to show the device image in the drop down menu. So copy it
from GTK+, clean it up, and keep using it.

Add a tool to test GcrImportButton