Makefile.am: Single Makefile mode and parallel-tests
gtk-doc doesn't work as an included Makefile so continue to
use SUBDIRS with docs/reference/*

Run the tests using TAP. Unfortunately the GTest --tap output
mode doesn't behave well in the face of failures, and doesn't
output a test plan. Use our own tap-driver and tap-compiler
for these reasons.

This also fixes several builddir != srcdir issues and testing hacks
that were sneaking around in the code base.

Move the tests into the same directories as the code that it
is testing.