Add (skip) annotation to gck_object_cache_get_attributes
This works around a bug in vala 0.27.1/git master where it would call
this method in the implementation of Pkcs11Certificate::set_property in
seahorse. This causes an infinite loop as
Pkcs11Certificate is an implementation of the GckObjectCacheInterface,
so calling gck_object_cache_get_attributes() for ::set_property which in
turn calls g_object_set(self, "attributes", value, NULL); is going to
loop.

This can be reverted once vala is fixed.

https://bugzilla.gnome.org/show_bug.cgi?id=745629