build: Fix GIR namespace version
Unless --nsversion is passed to g-ir-scanner, it will extract the
namespace version for a GIR file from the filename. Unfortunately, for
the Gcr-3.broken.gir file, this means it extracts ‘3.broken’. This is
incorrect, and doesn’t match the version in the typelib (which is the
correct ‘3’).

Having mismatched versions can cause problems when loading the GIR
files, as libgirepository complains that they don’t match the typelib.

https://bugzilla.gnome.org/show_bug.cgi?id=756235