Increase test-gnupg-collection timeout
The timeout length in the current test is too low and can make the test
fail if the system is too slow or too loaded. This test correctly stops
waiting when a result is received, so the timeout will not actually be
reached unless there is a problem.

[smcv: Expand commit message]
Fixes: https://gitlab.gnome.org/GNOME/gcr/issues/29
Bug-Debian: https://bugs.debian.org/846728
Signed-off-by: Simon McVittie <smcv@debian.org>