Update gcr_pkcs11_get_trust_{store|lookup}_slot URI checks
No store/lookup URIs may be present, even after initialising them, if no
appropriate stores are found. Check this is the case before attempting
to use them, and clarify in both debug output and the gtkdocs about what
is happening in such cases.

This makes the underlying problem in #10 a bit more obvious when
debugging.