build: Fix gcr-trust symbols not appearing in GIR, and hence also VAPI
Despite making it into the API docs, the trust APIs are not included
in the GCR GIR or VAPI files, since they were marked as private in the
Makefile, not public. As a result, apps using language bindings cannot
use the trust API without manual intervention.

This fixes the missing API by marking gcr-trust.c and gcr-trust.h
public.