build: Point POTFILES to the right files
The desktop files are generated in two passes:

 - the first, by configure taking the desktop.in.in template from
   the srcdir and replacing autoconf variables to generate a
   desktop.in file in the builddir
 - the second, by the Makefile taking the translations in the PO
   files from the srcdir, and running msgfmt on the desktop.in file
   in the builddir to generate a desktop file in the builddir

The xgettext utility can only extract strings from a file in the sources
directory, which means that the POTFILES file must contain a list of
files relative to the srcdir.

Since the translatable strings to be extracted are, by necessity, not
subject to substitution via autoconf/configure, we can safely point
xgettext to the desktop.in.in template file.

This change is necessary to make gcr build under non-srcdir build
conditions.

Signed-off-by: Emmanuele Bassi <ebassi@gnome.org>