build: Split desktop file generation rules
You cannot use the Makefile automatic variable $< for rules that involve
multiple files to mean "the dependency that triggered the rule". The $<
automatic variable means:

  The name of the first prerequisite.

So this:

    targets_in = foo.desktop.in bar.desktop.in
    targets = $(targets_in:.desktop.in=.desktop)
    $(targets): $(targets_in)
            do-something -o $@ $<

Will *not* expand to two rules:

    foo.desktop: foo.desktop.in
            do-something -o foo.desktop foo.desktop.in
    bar.desktop: bar.desktop.in
            do-something -o bar.desktop bar.desktop.in

It will, instead, expand to:

    foo.desktop: foo.desktop.in
            do-something -o foo.desktop foo.desktop.in
    bar.desktop: bar.desktop.in
            do-something -o bar.desktop foo.desktop.in

Which is definitely not what we want.

There is no automatic variable in Make that matches the Nth target to
the Nth dependency, which means we need to add an explicit rule for each
file.