Remove SKS network from keyserver defaults
The keyservers from the SKS network were recently infected with poisoned
certificates, as proclaimed here
https://gist.github.com/rjhansen/67ab921ffb4084c865b3618d6955275f.

As a result of importing these certificates, some people can experience
slow (to the point of being unusable) GPG tools. This is something that
cannot be "fixed" in SKS keyservers, so let's at least drop them from
the default keyservers we have in gcr (which are also used by seahorse).