Replace tap-gtester with one that relies on GLib 2.38+ TAP output
GLib 2.38 has been a hard dependency since gcr 3.13.91 (2014) so there
is little point in trying to accommodate older versions.

This resolves build-time test failures when built with GLib 2.61+,
where --tap is the default. The previous (Python) implementation of
tap-gtester attempted to auto-detect whether the output was TAP or
GTester, but detected GLib 2.61+ as GTester.

Signed-off-by: Simon McVittie <smcv@debian.org>
Fixes: https://gitlab.gnome.org/GNOME/gcr/issues/26