gck: Avoid unaligned memory access
When a value in GckAttribute is internally allocated, it embeds a
refcount (gint) at the beginning the allocated memory.  If refcount
has a smaller size than the actual value type (e.g., CK_ULONG), it
causes unaligned memory access when reading the value.

This patch allocates the maximum alignment size for the refcount part
so that the following value should always aligned.

See merge request GNOME/gcr!26.