gcr-certificate: Add (virtual) annotation
By adding `(virtual)` to a method, you tell the GObject Introspection
scanner that it is the public API of a certain vfunc. Since you can't
annotate a vfunc, GIR has a workaround where it will take over the
annotations of the invoker.

In other words, to make sure the annotations of the return value of
`GcrCertificate.get_der_data()` are understoor correctly, we just need
to add the necessary tags to `gcr_certificate_get_der_data()` and mark
it as `(virtual get_der_data)`.

Fixes https://gitlab.gnome.org/GNOME/gcr/issues/37