gck-slot: Improve assertion checks
Make more use of g_assert_cmpint and friends so that the error message
indicates more clearly what we were expecting and what we got.

This also has the minor benefit that unlike g_assert(), the more
test-oriented macros aren't disabled by G_DISABLE_ASSERT, so we don't
need to worry about whether we are relying on their side-effects.

Helps: #42
Signed-off-by: Simon McVittie <smcv@debian.org>