gck-call: Simplify the code by using GTask based implementation
GckCall uses its own implementation of threads pool to handle the async
calls, now that we've GTask this code can be simplified by reusing
GLib code.

I didn't want to change the API in this commit, even if private not to
mix changes together, so the functions still are used as they used to be

The main difference is that the async_pre and async_ready prepare a
GTask instance, while the async_go starts a thread using the gtask
function.

Callback functions needed to be adapted to the new GAsyncResult type.

Tests needed also some tuning as the underneath task holding a reference
to the source object might be finalized at later stage, when the
thread-related data is removed, as per this we may wait a bit to check
whether a source object gets finalized, this is due to GNOME/GLib#1346.