test-certificate-chain: Wait for objects being released by GTask's
As per switching to GTask we are facing GNOME/glib#1346 in various
places, including the certificate tests, as we were checking whether the
mock certificates were finalized in the same thread in which we created
them.

This may not always happen right now because GTask may release the last
reference of it later than we do.
So, to force this to happen let's wait until we're owing last reference
and only at that point we actually release it, so that we can be sure
that the objects are finalized on tear-down.

Fixes: https://gitlab.gnome.org/GNOME/gcr/-/issues/84