Drop custom icons in gcr
Icons are problematic to provide within gcr, as different DEs that are
using it, have different expectations what such an icon should look
like: one wants it to be colorful, another wants it to be monochrome,
yet another wants everything skeumorphic etc.

Let's not deal with this in gcr4 and drop any icons. If applications
want to provide an icon, they can do that themselves much better then we
ever can.