Support GnuTLS as an alternative crypto backend
This turns the `-Dgcrypt` build time option into a combo option, which
enables user to choose which cryptographic library to link with.  It
currently supports libgcrypt (`libgcrypt`) and GnuTLS (`gnutls`); for
the latter, a couple of legacy algorithms are disabled, namely
RIPEMD160 and PBES1-DES-SHA1 though it will be added in GnuTLS 3.8.5
for compatibility.

Signed-off-by: Daiki Ueno <dueno@src.gnome.org>